requirements-txt.txt
llama-index>=0.9.0
gradio>=4.0.0
pandas>=2.0.0
plotly>=5.18.0
python-docx>=0.8.11
fpdf>=1.7.2
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
kaleido  # for plotly static image export
python-dotenv>=0.19.0
